created CpuUtil so that x86 specific code can be kept out of FasterCode for arm64 compilation
