+++ bin/_fastercode/runVC32.bat
@@ -17,3 +17,7 @@ copy /B Faster_Irina.pyx+Faster_Polyglot.pyx FasterCode.pyx
 %FOLDER_PYTHON3%\python setup.py build_ext --inplace -i clean
 
 copy FasterCode.cp37-win32.pyd ..\..\OS\win32
+
+%FOLDER_PYTHON3%\python setup_cpuutil.py build_ext --inplace -i clean
+
+copy CpuUtil.cp37-win32.pyd ..\..\OS\win32
