+++ bin/_fastercode/source/Faster_Irina.pyx
@@ -115,10 +115,6 @@ cdef extern from "irina.h":
     void close_poly()
     void set_ext_fen_body(char * ext_fen, char * ext_body, char * pv )
 
-
-def bmi2():
-    return is_bmi2()
-
 class PGNreader:
     def __init__(self, path_file, depth):
         # fich
